Create MapActivity, move the map cam to Lleida and add a marker
