+++ app/src/main/java/predigsystem/udl/org/predigsystem/Activities/HomeActivity.java
@@ -23,7 +23,6 @@ import predigsystem.udl.org.predigsystem.Fragments.ArticleFragment;
 import predigsystem.udl.org.predigsystem.Fragments.CalendarFragment;
 import predigsystem.udl.org.predigsystem.Fragments.DashboardFragment;
 import predigsystem.udl.org.predigsystem.Fragments.HomeFragment;
-import predigsystem.udl.org.predigsystem.Fragments.MapFragment;
 import predigsystem.udl.org.predigsystem.Fragments.SettingsFragment;
 import predigsystem.udl.org.predigsystem.R;
 
@@ -97,10 +96,8 @@ public class HomeActivity extends AppCompatActivity{
                                         .commit();
                                 break;
                             case 2:
-                                fragment = new MapFragment();
-                                getSupportFragmentManager().beginTransaction()
-                                        .replace(R.id.content_layout, fragment)
-                                        .commit();
+                                Intent intentMap = new Intent(getApplicationContext(), MapActivity.class);
+                                startActivity(intentMap);
                                 break;
                             case 3:
                                 fragment = new CalendarFragment();
