+++ app/src/main/java/predigsystem/udl/org/predigsystem/Activities/MapActivity.java
@@ -1,15 +1,13 @@
-package predigsystem.udl.org.predigsystem.Fragments;
-
+package predigsystem.udl.org.predigsystem.Activities;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
+import android.support.v7.app.AppCompatActivity;
+import android.widget.Toast;
 
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
@@ -17,24 +15,22 @@ import com.google.android.gms.maps.model.MarkerOptions;
 import predigsystem.udl.org.predigsystem.R;
 
 
-/**
- * A simple {@link Fragment} subclass.
- */
-public class MapFragment extends Fragment implements OnMapReadyCallback{
+public class MapActivity extends AppCompatActivity implements OnMapReadyCallback {
 
     private GoogleMap mMap;
-    private static final LatLng CENTRECAT = new LatLng(41.837555, 1.537764);
 
-    public MapFragment() {
-        // Required empty public constructor
-    }
+    private static final LatLng CENTRECAT = new LatLng(41.837555, 1.537764);
+    private static final LatLng LLEIDA = new LatLng(41.607644, 0.622699);
+
 
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_map, container, false);
+    public void onCreate (Bundle savecInstanceBundle) {
+        super.onCreate(savecInstanceBundle);
+        setContentView(R.layout.fragment_map);
+
+        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
+        mapFragment.getMapAsync(this);
     }
 
     @Override
@@ -42,9 +38,8 @@ public class MapFragment extends Fragment implements OnMapReadyCallback{
         mMap = googleMap;
 
         // Add a marker in Lleida and move the camera
-        LatLng lleida = new LatLng(41, 0.62);
+        LatLng lleida = LLEIDA;
         mMap.addMarker(new MarkerOptions().position(lleida).title("Marker in Lleida").icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_RED)));
-        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(CENTRECAT, 7f));
+        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(LLEIDA, 12f));
     }
-
 }
