+++ app/src/main/res/layout/fragment_map.xml
@@ -4,7 +4,7 @@ android:layout_width="match_parent"
 android:layout_height="match_parent"
 xmlns:app="http://schemas.android.com/apk/res-auto"
 android:orientation="vertical"
-tools:context="predigsystem.udl.org.predigsystem.Fragments.MapFragment">
+tools:context="predigsystem.udl.org.predigsystem.Activities.MapActivity">
 
 <TextView
     android:id="@+id/textView"
